Avoid some float / double conversions in class Classify

+++ src/classify/normmatch.cpp
@@ -61,7 +61,7 @@ struct NORM_PROTOS {
  * normalization adjustment.  The equation that represents the transform is:
  *       1 / (1 + (NormAdj / midpoint) ^ curl)
  */
-static double NormEvidenceOf(double NormAdj) {
+static float NormEvidenceOf(float NormAdj) {
   NormAdj /= classify_norm_adj_midpoint;
 
   if (classify_norm_adj_curl == 3) {
@@ -69,9 +69,9 @@ static double NormEvidenceOf(double NormAdj) {
   } else if (classify_norm_adj_curl == 2) {
     NormAdj = NormAdj * NormAdj;
   } else {
-    NormAdj = pow(NormAdj, classify_norm_adj_curl);
+    NormAdj = pow(NormAdj, static_cast<float>(classify_norm_adj_curl));
   }
-  return (1.0 / (1.0 + NormAdj));
+  return (1 / (1 + NormAdj));
 }
 
 /*----------------------------------------------------------------------------
@@ -115,7 +115,7 @@ float Classify::ComputeNormMatch(CLASS_ID ClassId, const FEATURE_STRUCT &feature
     float Match = (feature.Params[CharNormLength] * feature.Params[CharNormLength] * 500.0f +
                    feature.Params[CharNormRx] * feature.Params[CharNormRx] * 8000.0f +
                    feature.Params[CharNormRy] * feature.Params[CharNormRy] * 8000.0f);
-    return (1.0f - NormEvidenceOf(Match));
+    return (1 - NormEvidenceOf(Match));
   }
 
   float BestMatch = FLT_MAX;
@@ -158,7 +158,7 @@ float Classify::ComputeNormMatch(CLASS_ID ClassId, const FEATURE_STRUCT &feature
       BestMatch = Match;
     }
   }
-  return 1.0 - NormEvidenceOf(BestMatch);
+  return 1 - NormEvidenceOf(BestMatch);
 } /* ComputeNormMatch */
 
 void Classify::FreeNormProtos() {
